move new exceptions into the api package
